+++ LogInWindow.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using DBCats.Tables;
+using DBCats;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -11,6 +13,8 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using MaterialDesignThemes.Wpf;
+using Microsoft.EntityFrameworkCore;
 
 namespace Cats_Program
 {
@@ -19,13 +23,43 @@ namespace Cats_Program
     /// </summary>
     public partial class LogInWindow : Window
     {
+        private SnackbarMessageQueue messageQueue = new SnackbarMessageQueue();
+
         public LogInWindow()
         {
             InitializeComponent();
+            Snackbar.MessageQueue = messageQueue;
         }
 
-        private void LogIn_Click(object sender, RoutedEventArgs e)
+        private async void LogIn_Click(object sender, RoutedEventArgs e)
         {
+            string login = Login.Text;
+            string password = Password.Password;
+
+            try
+            {
+                bool isUserAuthenticated = await Authenticate(login, password);
+
+                if (isUserAuthenticated)
+                {
+                    // Успешная аутентификация, выполните действия, которые вам необходимы
+                    ShowSnackbar("Успешный вход!");
+                    CatWindow catWindow = new CatWindow();
+                    catWindow.Show();
+                    Close();
+
+                }
+                else
+                {
+                    // Пользователь не найден, выполните действия по обработке ошибки
+                    ShowSnackbar("Ошибка аутентификации");
+                }
+            }
+            catch (Exception ex)
+            {
+                // Обработка ошибок базы данных
+                ShowSnackbar("Произошла ошибка: " + ex.Message);
+            }
 
         }
 
@@ -35,5 +69,20 @@ namespace Cats_Program
             signup.Show();
             Close();
         }
+        private void ShowSnackbar(string message)
+        {
+            messageQueue.Enqueue(message);
+        }
+        private async Task<bool> Authenticate(string login, string password)
+        {
+            using (var context = new CatsDBContext())
+            {
+                // Проверяем, существует ли пользователь с указанным логином и паролем
+                var user = await context.User.SingleOrDefaultAsync(u => u.Login == login && u.Password == password);
+
+                return user != null;
+            }
+        }
+
     }
 }
